remove old components page
