--- packages/commonwealth/client/scripts/views/pages/components.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { ComponentShowcase } from '../components/component_kit/cw_component_showcase';
-
-const ComponentsPage = () => {
-  return <ComponentShowcase />;
-};
-
-export default ComponentsPage;
